+++ packages/aws-lambda/index.js
@@ -1,5 +1,6 @@
 /* eslint-disable global-require */
-if (process.env.BUILD) {
+
+if (process.env.WEBPACK_BUILD) {
   module.exports = require('./irving.config.server.js');
 }
 /* eslint-enable */