+++ packages/aws-lambda/irving.config.server.js
@@ -1,0 +1,6 @@
+const awsLambdaConfig = require('./config/awsLambdaConfig');
+
+module.exports = {
+  name: 'aws-lambda',
+  exportServer: awsLambdaConfig,
+};