+++ packages/core/cli/start.js
@@ -1,6 +1,7 @@
 const {
   rootUrl
 } = require('../config/paths');
+
 const getConfigField = require('../utils/getConfigField');
 const startServer = require('../server/startServer');
 const app = require('../server');