+++ packages/core/server/index.js
@@ -3,7 +3,7 @@
 const getEnv = require('../config/env');
 const {
   API_ROOT_URL,
-  API_ORIGN,
+  API_ORIGIN,
   NODE_ENV,
 } = getEnv();
 
@@ -15,8 +15,9 @@ getService().start();
 require('../utils/shimWindow');
 
 const express = require('express');
-const bodyParser = require('body-parser');
-const proxy = require('http-proxy-middleware');
+const {
+  createProxyMiddleware
+} = require('http-proxy-middleware');
 const cookiesMiddleware = require('universal-cookie-express');
 const getConfigField = require('../utils/getConfigField');
 const {
@@ -25,15 +26,11 @@ const {
 
 const getLogService = require('../services/logService');
 const customizeRedirect = require('../server/customizeRedirect');
-const purgeCache = require('../server/purgeCache');
-const getCacheKeys = require('../server/getCacheKeys');
-
 const log = getLogService('irving:server');
 const app = express();
 
-// Clearing the Redis cache.
-app.post('/purge-cache', bodyParser.json(), purgeCache);
-app.get('/cache-keys', getCacheKeys);
+// Cache-related endpoints.
+// require('../server/cache')(app);
 
 // Set view engine.
 app.set('view engine', 'ejs');
@@ -42,27 +39,30 @@ app.set('view engine', 'ejs');
 const irvingServerMiddleware = getConfigField('customizeServer');
 irvingServerMiddleware.forEach((middleware) => middleware(app));
 
+// const renato = getConfigField('exportServer');
+// console.log(renato);
+
 // Set up a reusable proxy for responses that should be served directly.
-const proxyPassthrough = getConfigArray('proxyPassthrough');
-const passthrough = proxy({
-  changeOrigin: true,
-  followRedirects: true,
-  secure: 'development' !== NODE_ENV,
-  // @todo make this not specific to WP eventually.
-  target: API_ORIGN || API_ROOT_URL.replace('/wp-json/irving/v1', ''),
-  xfwd: true,
-});
+// const proxyPassthrough = getConfigArray('proxyPassthrough');
+// const passthrough = createProxyMiddleware({
+//   changeOrigin: true,
+//   followRedirects: true,
+//   secure: 'development' !== NODE_ENV,
+//   // @todo make this not specific to WP eventually.
+//   target: API_ORIGIN || API_ROOT_URL.replace('/wp-json/irving/v1', ''),
+//   xfwd: true,
+// });
 
 // Create proxies for each configured proxy pattern.
-proxyPassthrough.forEach((pattern) => {
-  app.use(pattern, passthrough);
-});
+// proxyPassthrough.forEach((pattern) => {
+//   app.use(pattern, passthrough);
+// });
 
 // Add universal cookies middleware.
 app.use(cookiesMiddleware());
 
 // Naked Redirect.
-app.use(customizeRedirect());
+// app.use(customizeRedirect());
 
 if ('development' === NODE_ENV) {
   require('../server/development')(app);