+++ packages/lambda/irving.config.server.js
@@ -1,6 +1,0 @@
-const exportServer = require('./config/exportServer.js');
-
-module.exports = {
-  name: 'lambda',
-  exportServer,
-};